Server/src/requirements.txt
asgiref==3.7.2
Django==4.2.4
Pillow==10.0.0
python-decouple==3.8
djangorestframework~=3.14.0